+++ Changes.md
@@ -23,12 +23,18 @@ Improvements
 - AttributeTweaks, CameraTweaks, ShaderTweaks, OptionTweaks, PrimitiveVariableTweaks :
   - Added support for a `{source}` token which is substituted with the original value when tweaking a string in `Replace` mode.
   - Added tooltips documenting the tweak modes.
+- 3Delight : Added automatic render-time translation of UsdPreviewSurface shaders to 3Delight.
+- USDLight :
+  - Added Arnold-specific extension parameters.
+  - Added parameter tooltips.
+- LightEditor : Added columns for Arnold-specific parameters on USD lights.
 
 Fixes
 -----
 
 - Viewer : Fixed Cycles shader balls.
 - TweakPlug : Fixed incorrect results and potential crashes in list modes.
+- USD : Fixed `Unsupported value type "StringData" for parameter "input"` warning when converting `UsdTransform2d` shaders with no `in` connections to Arnold.
 
 API
 ---
@@ -37,6 +43,8 @@ API
 - BoxPlug : Added Python bindings for `ValueType`, `PointType` and `ChildType` type aliases.
 - RenderPassEditor : Added `deregisterColumn()` method.
 - DocumentationAlgo : Added table and strikethrough support to `markdownToHTML()`.
+- LightEditor : Added `columnName` parameter to `registerParameter()` method, matching the behaviour of `RenderPassEditor.registerOption()`.
+- USDShader : Added support for loading from the UsdSchemaRegistry as well as from the SdrRegistry. This is now used when loading UsdLuxLights.
 
 Build
 -----